Vecs & MoveSemantics

+++ exercises/01_variables/variables6.rs
@@ -3,7 +3,7 @@
 // Execute `rustlings hint variables6` or use the `hint` watch subcommand for a
 // hint.
 
-const NUMBER: usize = 3; // when we define a constant, we need to give a type declaration. doesn't infer type
+const NUMBER: u32 = 3; // when we define a constant, we need to give a type declaration. doesn't infer type
 fn main() {
     println!("Number {}", NUMBER);
 }
